(MAINT) Convert status state string to kwd
TK status requires the state value in map to be a keyword like
`:running`. When this is serialized to JSON and sent over the wire (like
would happen with a status function call) that string needs to be
converted back into a keyword. This ensures local callers (i.e. users of
the RbacConsumerService in `pe-rbac-service`) and remote
callers (i.e. users of this library) will get consistent results.