(PDB-2853) Return state-map when rbac-url configuration is missing
This commit returns `{:state :unknown}` instead of `:unknown` when the
rbac url is missing in the user configuration. Prior to this commit
consumers of the status function would call `:state` on the return value
and if the configuration was missing they would get `null` and add that
as the status of rbac which is not consistent with the tk-status api.