(PE-16563) Add middleware for cert authentication
This adds a wrap-token-and-cert-access middleware which applies both
token and cert-based authentication to a ring handler. The cert auth
assumes that the client certificate on a request has been validated by
the webserver and is thus trusted by the CA. It then uses a new
cert->subject protocol function to authenticate the user based on the
CN. The cert->subject function will return the subject associated with
the cert or nil if none is. If it returns a subject, that subject will
be attached to the request.

For now, the subject will always be the api_user if the cert is on the
whitelist and will otherwise by nil, but it's possible the future may
include other users associated with certs.