(maint) change are-permitted? to always return a vector
Presently the dummy version of are-permitted? returns the boolean true, and the
real version returns false in the case of an improperly configured
rbac-consumer. This is a little dangerous because the method is documented as
returning a list of booleans, and returning an actual boolean in
misconfiguration cases can lead to uncaught bugs.

This commit changes the dummy implementation to return a boolean vector, and
changes the misconfiguration case to throw an exception on init, rather than
logging an error and returning false.