(PE-18964) Accept no-keepalive tokens.
Allow the token passed to the `valid-token->subject` method of the
RbacConsumerService protocol to be suffixed with `|no_keepalive`. If
present, this suffix causes the consumer service to avoid updating the
token's last_active time, by setting the `update_last_activity?` field
to `false` in the body it submits to RBAC's v2 token auth endpoint.