(PE-23159) adds list-permitted interface to protocol and client
This adds a new function "list-permitted" that given a token,
an object type, and an action, returns the list of permissions
a user has access to.

The client is updated to include an uncertified client for
making this request as we want the authentication to be made through
the token that is provided, and not the ssl certificates.