(maint) use URI, update clj-parent, update certs
The certs used in the test environment expired in January.  This
updates them to more recent versions.

The clj-parent version used in the test environment was very old,
so this updates it to the most recent version.

java.net.URI is preferred over java.net.URL as it doesn't require
a DNS lookup, so the use of that was replaced.