(maint) Handle missing cert in cert-related functions
Previously, the `cert-whitelisted?` and `cert->subject` functions would
generate an incorrect URL when called without a cert. This commit
updates them to return nil in that case, which matches the behavior of
the standard implementations in the RBAC service.

This case is important because the cert-based authentication middleware
in RBAC calls cert->subject regardless of whether a cert was actually
supplied, relying on the function to return nil when no cert is present.
This bug causes issues for services depending on the cert middleware
in cases where they're running with a remote RBAC service.