(PE-29720) add support for activity service v2 POST endpoint
This adds support for the upcoming v2 endpoint for the activity service
that allows multiple objects and an ip-address to be included in
the activity service submission.

It updates the changelog, updates the version number to 1.0 and
adds some simple endpoint testing for the client code.

The service initially attempts to post to the v2 endpoint.  If that
results in a 404, the service caches that knowledge and resubmits
to the v1 endpoint.  Subsequent requests will use the v1 endpoint
exclusively.