(PE-29996) deprecate cert-whitelisted?, use rbac v2 endpoint
This deprecates the cert-whitelisted? function in the rbac remote consumer interface.
It also adds a new function to the protocol for the rbac-consumer `cert-allowed?`
which mirrors the behavior of the old function with some small differences.  First
the new "v2" endpoint in RBAC is attempted to be used to resolve the validity
of the certificate.  If the v2 endpoint results in a 404 response, that is cached
and the v1 endpoint is returned.  Future requests will use the v1 endpoint if the
cached result shows that the v2 endpoint wasn't found.

Tests were updated to demonstrate the new endpoints and the behavior.