(maint) Update str->uuid function to allow parsing UUIDs
The str-uuid function is used in the parse-subject function to ensure subject ids are proper UUID objects. If an id is already a UUID instead of raising an exception it is helpful to just return that UUID instead of rasing a `ClassCastException`. The use case for this change is to use `parse-subject` to normalize subjects when using the rbac client, for example replacing `->rbac-subject` in https://github.com/puppetlabs/orchestrator/blob/c4d3e0da65eab03e887b07fe8ed13407014a388e/src/puppetlabs/orchestrator/rbac.clj#L10-L19 where subjects may or may not already have their IDs in UUID form based on whether or not they have been serialized/deserialized from for example a DB query.